Released 3 functions: 1. Register user 2. Register school 3. Register school class

+++ src/main/kotlin/com/school1c/kotlinschool/entities/SchoolClass.kt
@@ -5,7 +5,7 @@ import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Table
 
 @Table(value = "school_classes")
-data class SchoolClass(@Id val id:Long? = null, val name:String, private val school: Long) {
+data class SchoolClass(@Id val id:Int? = null, val name:String, private val school: Int) {
 //    suspend fun<T:User> getUsers(userRepository: UserRepository<T>): Flow<T> {
 //        return userRepository.findAllBySchoolClassId(id!!)
 //    }
